상품 조회 완료

+++ src/main/java/com/data/dataproject/dto/mypage/OrderDto.java
@@ -2,6 +2,6 @@ package com.data.dataproject.dto.mypage;
 
 public class OrderDto {
     private String productName;
-    private Integer productCount;
-    private Integer productPrice;
+    private Long productCount;
+    private Long productPrice;
 }
